# tsdb: Save allocations on labels by re-using label array