I believe this should fix https://github.com/prometheus/prometheus/issues/7732.

@bwplotka @codesome @brian-brazil