@krasi-georgiev 2020-08-04
there is an update for funcbench that should solve this. will be merged today