@roidelapluie 2020-08-04
Yes but it needs an update that will arrive today.