@bwplotka 2020-08-04
Yes @brancz it would be appreciated if someone has time to run `BenchmarkQuerierSelect` locally and compare with and without this PR :hugs: 

That's the only thing left (: There is a rule: **Never trust your eyes in terms of perf improvements until you see results** ;p