@brancz 2020-08-04
@bwplotka I meant the funcbench "fix" :)