@brancz 2020-08-04
Looks mixed:

```
benchmark                                                old ns/op      new ns/op      delta
BenchmarkQuerierSelect/Head/1of1000000-8                 466893848      380383208      -18.53%
BenchmarkQuerierSelect/Head/10of1000000-8                497267937      412377178      -17.07%
BenchmarkQuerierSelect/Head/100of1000000-8               407210963      394276352      -3.18%
BenchmarkQuerierSelect/Head/1000of1000000-8              392726174      386735189      -1.53%
BenchmarkQuerierSelect/Head/10000of1000000-8             396210322      420197867      +6.05%
BenchmarkQuerierSelect/Head/100000of1000000-8            428185499      470494794      +9.88%
BenchmarkQuerierSelect/Head/1000000of1000000-8           671177681      738573544      +10.04%
BenchmarkQuerierSelect/SortedHead/1of1000000-8           1775712879     2086305777     +17.49%
BenchmarkQuerierSelect/SortedHead/10of1000000-8          1735133140     1855856583     +6.96%
BenchmarkQuerierSelect/SortedHead/100of1000000-8         1798964578     1928457124     +7.20%
BenchmarkQuerierSelect/SortedHead/1000of1000000-8        1687464406     2082599465     +23.42%
BenchmarkQuerierSelect/SortedHead/10000of1000000-8       1681772608     2354247158     +39.99%
BenchmarkQuerierSelect/SortedHead/100000of1000000-8      1698929506     2027262925     +19.33%
BenchmarkQuerierSelect/SortedHead/1000000of1000000-8     1959577713     2138919743     +9.15%
BenchmarkQuerierSelect/Block/1of1000000-8                574979728      567111772      -1.37%
BenchmarkQuerierSelect/Block/10of1000000-8               583562770      577580506      -1.03%
BenchmarkQuerierSelect/Block/100of1000000-8              600302107      567982871      -5.38%
BenchmarkQuerierSelect/Block/1000of1000000-8             575752036      580213844      +0.77%
BenchmarkQuerierSelect/Block/10000of1000000-8            622611077      810807972      +30.23%
BenchmarkQuerierSelect/Block/100000of1000000-8           698878044      1016014379     +45.38%
BenchmarkQuerierSelect/Block/1000000of1000000-8          1128960288     1140317753     +1.01%
```