@krasi-georgiev 2020-08-04
> Are we actually calling funcbench correctly?

yes the issue is somewhere else, when you provide wrong args to funcbench it will post back a comment to notify about the wrong usage and also prints a link to the readme for the correct usage