@brancz 2020-08-05
Tried a few more things, latest benchmark is now:

```
○ benchcmp old.txt new.txt                                                                                             
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat                       
benchmark                                                old ns/op      new ns/op      delta    
BenchmarkQuerierSelect/Head/1of1000000-8                 296799961      285314344      -3.87%                          
BenchmarkQuerierSelect/Head/10of1000000-8                287250120      287423672      +0.06%   
BenchmarkQuerierSelect/Head/100of1000000-8               289651377      283307649      -2.19%                          
BenchmarkQuerierSelect/Head/1000of1000000-8              287464786      290702025      +1.13%   
BenchmarkQuerierSelect/Head/10000of1000000-8             289590086      288229907      -0.47%
BenchmarkQuerierSelect/Head/100000of1000000-8            307543377      302647535      -1.59%
BenchmarkQuerierSelect/Head/1000000of1000000-8           462260730      430338135      -6.91%
BenchmarkQuerierSelect/SortedHead/1of1000000-8           1101255421     1078921062     -2.03%
BenchmarkQuerierSelect/SortedHead/10of1000000-8          1103327856     1095498903     -0.71%
BenchmarkQuerierSelect/SortedHead/100of1000000-8         1088456364     1087725571     -0.07%
BenchmarkQuerierSelect/SortedHead/1000of1000000-8        1091687651     1098109147     +0.59%
BenchmarkQuerierSelect/SortedHead/10000of1000000-8       1092085332     1089343240     -0.25%
BenchmarkQuerierSelect/SortedHead/100000of1000000-8      1116357490     1113042790     -0.30%
BenchmarkQuerierSelect/SortedHead/1000000of1000000-8     1255759984     1215396251     -3.21%
BenchmarkQuerierSelect/Block/1of1000000-8                501893927      489486589      -2.47%
BenchmarkQuerierSelect/Block/10of1000000-8               500186117      486818372      -2.67%
BenchmarkQuerierSelect/Block/100of1000000-8              508082676      499404936      -1.71%
BenchmarkQuerierSelect/Block/1000of1000000-8             502062640      500199044      -0.37%
BenchmarkQuerierSelect/Block/10000of1000000-8            512914331      500545053      -2.41%
BenchmarkQuerierSelect/Block/100000of1000000-8           536542749      513072876      -4.37% 
BenchmarkQuerierSelect/Block/1000000of1000000-8          719338478      616495805      -14.30%
```

tl;dr huge queries get faster, small queries stay roughly the same