@geekodour 2020-08-05
/funcbench master BenchmarkQuerierSelect

hoping this works as expected! :pray: