@brancz 2020-08-05
I don't understand the funcbench output, when literally running the same command gives me this:

```
benchmark                                                old ns/op      new ns/op      delta
BenchmarkQuerierSelect/Head/1of1000000-8                 292395617      280572216      -4.04%
BenchmarkQuerierSelect/Head/10of1000000-8                288081239      288336309      +0.09%
BenchmarkQuerierSelect/Head/100of1000000-8               293925265      281808322      -4.12%
BenchmarkQuerierSelect/Head/1000of1000000-8              294774636      290407317      -1.48%
BenchmarkQuerierSelect/Head/10000of1000000-8             292633875      286288476      -2.17%
BenchmarkQuerierSelect/Head/100000of1000000-8            306918348      301298413      -1.83%
BenchmarkQuerierSelect/Head/1000000of1000000-8           463544980      423725674      -8.59%
BenchmarkQuerierSelect/SortedHead/1of1000000-8           1126705413     1099482013     -2.42%
BenchmarkQuerierSelect/SortedHead/10of1000000-8          1126151936     1113746536     -1.10%
BenchmarkQuerierSelect/SortedHead/100of1000000-8         1117822136     1108380530     -0.84%
BenchmarkQuerierSelect/SortedHead/1000of1000000-8        1131196263     1103403956     -2.46%
BenchmarkQuerierSelect/SortedHead/10000of1000000-8       1117592481     1112906157     -0.42%
BenchmarkQuerierSelect/SortedHead/100000of1000000-8      1136235830     1111508887     -2.18%
BenchmarkQuerierSelect/SortedHead/1000000of1000000-8     1280433034     1228725261     -4.04%
BenchmarkQuerierSelect/Block/1of1000000-8                513837151      531370128      +3.41%
BenchmarkQuerierSelect/Block/10of1000000-8               514763484      525483317      +2.08%
BenchmarkQuerierSelect/Block/100of1000000-8              518187861      529311642      +2.15%
BenchmarkQuerierSelect/Block/1000of1000000-8             521457922      526518763      +0.97%
BenchmarkQuerierSelect/Block/10000of1000000-8            529431477      513692094      -2.97%
BenchmarkQuerierSelect/Block/100000of1000000-8           545169522      532103644      -2.40%
BenchmarkQuerierSelect/Block/1000000of1000000-8          745794602      664299538      -10.93%

benchmark                                                old allocs     new allocs     delta
BenchmarkQuerierSelect/Head/1of1000000-8                 1000015        1000013        -0.00%
BenchmarkQuerierSelect/Head/10of1000000-8                1000051        1000040        -0.00%
BenchmarkQuerierSelect/Head/100of1000000-8               1000411        1000310        -0.01%
BenchmarkQuerierSelect/Head/1000of1000000-8              1004011        1003010        -0.10%
BenchmarkQuerierSelect/Head/10000of1000000-8             1040011        1030010        -0.96%
BenchmarkQuerierSelect/Head/100000of1000000-8            1400011        1300010        -7.14%
BenchmarkQuerierSelect/Head/1000000of1000000-8           5000010        4000010        -20.00%
BenchmarkQuerierSelect/SortedHead/1of1000000-8           1000052        1000050        -0.00%
BenchmarkQuerierSelect/SortedHead/10of1000000-8          1000087        1000077        -0.00%
BenchmarkQuerierSelect/SortedHead/100of1000000-8         1000447        1000347        -0.01%
BenchmarkQuerierSelect/SortedHead/1000of1000000-8        1004047        1003047        -0.10%
BenchmarkQuerierSelect/SortedHead/10000of1000000-8       1040047        1030047        -0.96%
BenchmarkQuerierSelect/SortedHead/100000of1000000-8      1400047        1300047        -7.14%
BenchmarkQuerierSelect/SortedHead/1000000of1000000-8     5000047        4000047        -20.00%
BenchmarkQuerierSelect/Block/1of1000000-8                3000016        3000012        -0.00%
BenchmarkQuerierSelect/Block/10of1000000-8               3000050        3000030        -0.00%
BenchmarkQuerierSelect/Block/100of1000000-8              3000410        3000210        -0.01%
BenchmarkQuerierSelect/Block/1000of1000000-8             3004010        3002010        -0.07%
BenchmarkQuerierSelect/Block/10000of1000000-8            3040010        3020010        -0.66%
BenchmarkQuerierSelect/Block/100000of1000000-8           3400010        3200010        -5.88%
BenchmarkQuerierSelect/Block/1000000of1000000-8          7000010        5000010        -28.57%

benchmark                                                old bytes     new bytes     delta
BenchmarkQuerierSelect/Head/1of1000000-8                 48000784      48000992      +0.00%
BenchmarkQuerierSelect/Head/10of1000000-8                48002368      48002288      -0.00%
BenchmarkQuerierSelect/Head/100of1000000-8               48018208      48015248      -0.01%
BenchmarkQuerierSelect/Head/1000of1000000-8              48176608      48144848      -0.07%
BenchmarkQuerierSelect/Head/10000of1000000-8             49760608      49440848      -0.64%
BenchmarkQuerierSelect/Head/100000of1000000-8            65600608      62400848      -4.88%
BenchmarkQuerierSelect/Head/1000000of1000000-8           224000546     192000850     -14.29%
BenchmarkQuerierSelect/SortedHead/1of1000000-8           101191800     101192008     +0.00%
BenchmarkQuerierSelect/SortedHead/10of1000000-8          101193320     101193304     -0.00%
BenchmarkQuerierSelect/SortedHead/100of1000000-8         101209160     101206264     -0.00%
BenchmarkQuerierSelect/SortedHead/1000of1000000-8        101367560     101335864     -0.03%
BenchmarkQuerierSelect/SortedHead/10000of1000000-8       102951560     102631864     -0.31%
BenchmarkQuerierSelect/SortedHead/100000of1000000-8      118791560     115591864     -2.69%
BenchmarkQuerierSelect/SortedHead/1000000of1000000-8     277191560     245191864     -11.54%
BenchmarkQuerierSelect/Block/1of1000000-8                99200840      99200984      +0.00%
BenchmarkQuerierSelect/Block/10of1000000-8               99202472      99202136      -0.00%
BenchmarkQuerierSelect/Block/100of1000000-8              99219752      99213656      -0.01%
BenchmarkQuerierSelect/Block/1000of1000000-8             99392552      99328856      -0.06%
BenchmarkQuerierSelect/Block/10000of1000000-8            101120552     100480856     -0.63%
BenchmarkQuerierSelect/Block/100000of1000000-8           118400552     112000856     -5.41%
BenchmarkQuerierSelect/Block/1000000of1000000-8          291200552     227200856     -21.98%
```